// src/reducers/result/result.reducer.types.ts
export type InitialResultState = {
  attemptedQuestions: number;
  rightAnswers: number;
  wrongAnswers: number;
};

export type ResultAction =
  | { type: "ATTEMPTED_QUESTIONS"; payload: { questions: number } }
  | { type: "RIGHT_ANSWERS"; payload: { rightAnswers: number } }
  | { type: "WRONG_ANSWERS"; payload: { wrongAnswers: number } };